Move `getFontDataFromUrl` to `wp-dashboard`
