+++ packages/wp-dashboard/src/components/editorSettings/customFonts/utils/getFontDataFromUrl.js
